Move AdapterInfo into wgt
